requirements-py313.txt
# Requisitos compatibles con Python 3.13.x
# Probado en 3.13.6

# Core Flask (usar versiones 3.x compatibles con 3.13)
Flask>=3.0.0,<4.0
flask-cors>=4.0.0
Flask-Caching>=2.1.0
Werkzeug>=3.0.0,<4.0

# Database
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
SQLAlchemy>=2.0.31

# AI & Processing
# google-generativeai 0.7+ funciona con 3.13
google-generativeai>=0.7.0
openai>=1.30.0
tenacity>=8.2.3
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Utilities
python-dotenv>=1.0.1
PyYAML>=6.0.1
structlog>=24.1.0

# Document processing
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2

# Monitoring (opcional)
sentry-sdk[flask]>=2.0.0

# Production server
# gunicorn no aplica en Windows, usar waitress o Flask dev server.
# Para Linux/WSGI:
gunicorn>=23.0.0

# Development
pytest>=8.2.0
pytest-flask>=1.3.0
black>=24.4.0
flake8>=7.1.0

# General
setuptools>=70.0.0
